+++ src/test/java/model/piece/PieceMoveTest.java
@@ -11,8 +11,6 @@ import org.junit.jupiter.api.Test;
 import model.Position;
 import model.Team;
 import model.board.Board;
-import model.piece.iterable.Chariot;
-import model.piece.iterable.Pao;
 import model.piece.normal.Elephant;
 import model.piece.normal.Palace;
 import model.piece.normal.Pawn;
